@@ -9,7 +9,7 @@ use Doctrine\Persistence\ObjectManager;
 
 class FormFixtures extends Fixture implements FixtureGroupInterface
 {
-    private $fixturesHelpers;
+    private FixturesHelpers $fixturesHelpers;
 
     public function __construct(FixturesHelpers $fixturesHelpers)
     {
@@ -51,6 +51,6 @@ class FormFixtures extends Fixture implements FixtureGroupInterface
      */
     public static function getGroups(): array
     {
-        return ['component', 'component-form'];
+        return ['component', 'builder-components', 'component-form'];
     }
 }
