@@ -9,7 +9,7 @@ use Doctrine\Persistence\ObjectManager;
 
 class MapFixtures extends Fixture implements FixtureGroupInterface
 {
-    private $fixturesHelpers;
+    private FixturesHelpers $fixturesHelpers;
 
     public function __construct(FixturesHelpers $fixturesHelpers)
     {
@@ -69,6 +69,6 @@ class MapFixtures extends Fixture implements FixtureGroupInterface
      */
     public static function getGroups(): array
     {
-        return ['component','component-map'];
+        return ['component', 'builder-components','component-map'];
     }
 }
