@@ -9,7 +9,7 @@ use Doctrine\Persistence\ObjectManager;
 
 class RowFixtures extends Fixture implements FixtureGroupInterface
 {
-    private $fixturesHelpers;
+    private FixturesHelpers $fixturesHelpers;
 
     public function __construct(FixturesHelpers $fixturesHelpers)
     {
@@ -115,6 +115,6 @@ class RowFixtures extends Fixture implements FixtureGroupInterface
      */
     public static function getGroups(): array
     {
-        return ['component', 'component-row'];
+        return ['component', 'builder-components', 'component-row'];
     }
 }
