@@ -2,7 +2,6 @@
 
 namespace Akyos\BuilderBundle\Service;
 
-use Akyos\BuilderBundle\Controller\BuilderController;
 use Akyos\BuilderBundle\Entity\BuilderTemplate;
 use Akyos\BuilderBundle\Entity\Component;
 use Akyos\BuilderBundle\Entity\ComponentTemplate;
@@ -13,41 +12,38 @@ use Akyos\BuilderBundle\Form\MakeTemplateType;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\Query;
 use Doctrine\ORM\QueryBuilder;
-use Gedmo\Translatable\Entity\Translation;
 use Gedmo\Translatable\Query\TreeWalker\TranslationWalker;
 use Gedmo\Translatable\TranslatableListener;
+use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\ContainerInterface;
+use Psr\Container\NotFoundExceptionInterface;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\RequestStack;
-use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Twig\Environment;
+use Twig\Error\LoaderError;
+use Twig\Error\RuntimeError;
+use Twig\Error\SyntaxError;
 
 class Builder
 {
-    private $router;
-    /** @var BuilderController */
-    private $builderController;
-    /** @var Request|null */
-    private $request;
-    /** @var EntityManagerInterface */
-    private $em;
-    /** @var Environment */
-    private $environment;
-    /** @var ContainerInterface */
-    private $container;
+    private ?Request $request;
+    private EntityManagerInterface $em;
+    private Environment $environment;
+    private ContainerInterface $container;
 
-    public function __construct(UrlGeneratorInterface $router, BuilderController $builderController, RequestStack $requestStack, EntityManagerInterface $em, Environment $environment, ContainerInterface $container)
+    public function __construct(RequestStack $requestStack, EntityManagerInterface $em, Environment $environment, ContainerInterface $container)
     {
-        $this->router = $router;
-        $this->builderController = $builderController;
         $this->request = $requestStack->getCurrentRequest();
         $this->em = $em;
         $this->environment = $environment;
         $this->container = $container;
     }
 
-    public function getTab()
+    /**
+     * @return string
+     */
+    public function getTab(): string
     {
         $tab = '<li class="nav-item">';
             $tab .= '<a class="nav-link" id="builder-tab" data-toggle="tab" href="#builder" role="tab" aria-controls="builder" aria-selected="false">Builder</a>';
@@ -55,9 +51,18 @@ class Builder
         return $tab;
     }
 
+    /**
+     * @param $objectType
+     * @param $objectId
+     * @return string|RedirectResponse
+     * @throws ContainerExceptionInterface
+     * @throws NotFoundExceptionInterface
+     * @throws LoaderError
+     * @throws RuntimeError
+     * @throws SyntaxError
+     */
     public function getTabContent($objectType, $objectId)
     {
-//        ->getRepository(Component::class)->findBy(['type' => $objectType, 'typeId' => $objectId, 'isTemp' => true, 'parentComponent' => null], ['position' => 'ASC']);
         /** @var QueryBuilder $qbc */
         $qbc = $this->em->getRepository(Component::class)->createQueryBuilder('c');
         $qbc
@@ -116,7 +121,14 @@ class Builder
         ]);
     }
 
-    public function initCloneComponents($type, $typeId)
+    /**
+     * @param $type
+     * @param $typeId
+     * @return bool
+     * @throws ContainerExceptionInterface
+     * @throws NotFoundExceptionInterface
+     */
+    public function initCloneComponents($type, $typeId): bool
     {
         $components = $this->em->getRepository(Component::class)->findBy(array('type' => $type, 'typeId' => $typeId, 'parentComponent' => null));
         $componentsTemp = $this->em->getRepository(Component::class)->findBy(array('type' => $type, 'typeId' => $typeId, 'isTemp' => true));
@@ -136,24 +148,28 @@ class Builder
         // Check if there are already temp components
         if (!$componentsTemp) {
             foreach ($components as $component) {
-                if ($component instanceof Component) {
-                    // If he has childs
-                    if (!($component->getParentComponent())) {
-                        $this->cloneComponent($component);
-                    }
+                // If he has childs
+                if (($component instanceof Component) && !($component->getParentComponent())) {
+                    $this->cloneComponent($component);
                 }
             }
             $this->em->flush();
         }
+
+        return true;
     }
 
-    public function tempToProd($type, $typeId)
+    /**
+     * @param $type
+     * @param $typeId
+     * @return bool
+     */
+    public function tempToProd($type, $typeId): bool
     {
         // Get each components of page.
         $newComponents = $this->em->getRepository(Component::class)->findBy(array('type' => $type, 'typeId' => $typeId));
         $componentsProd = $this->em->getRepository(Component::class)->findBy(array('type' => $type, 'typeId' => $typeId, 'isTemp' => false));
         $componentsTemp = $this->em->getRepository(Component::class)->findBy(array('type' => $type, 'typeId' => $typeId, 'isTemp' => true));
-
 
         // TODO => If there are no more temp component.
         if ($componentsProd && !$componentsTemp) {
@@ -179,9 +195,20 @@ class Builder
 
             $this->em->flush();
         }
+
+        return true;
     }
 
-    public function cloneComponent(Component $component, $parent = null, $changeType = null, $changeTypeId = null, int $nextPos = null, $makeTemplate = false)
+    /**
+     * @param Component $component
+     * @param null $parent
+     * @param null $changeType
+     * @param null $changeTypeId
+     * @param int|null $nextPos
+     * @param false $makeTemplate
+     * @return bool
+     */
+    public function cloneComponent(Component $component, $parent = null, $changeType = null, $changeTypeId = null, int $nextPos = null, $makeTemplate = false): bool
     {
         $clone = clone $component;
         // Clone each component of page and set isTemp to true
@@ -193,7 +220,7 @@ class Builder
         if ($changeType && $changeTypeId) {
             $clone->setType($changeType);
             $clone->setTypeId($changeTypeId);
-            $clone->setPosition(intval($nextPos));
+            $clone->setPosition((int)$nextPos);
         }
         if ($parent && $parent instanceof Component) {
             $clone->setParentComponent($parent);
@@ -231,7 +258,7 @@ class Builder
         }
         if ($component->getChildComponents()) {
             foreach ($clone->getChildComponents() as $childComponent) {
-                if ($childComponent->getIsTemp() == false) {
+                if ($childComponent->getIsTemp() === false) {
                     $childComponent->setParentComponent($component);
                     $clone->removeChildComponent($childComponent);
                 }
@@ -241,17 +268,32 @@ class Builder
             }
         }
         $this->em->flush();
+
+        return true;
     }
 
-    public function onDeleteEntity($type, $typeId)
+    /**
+     * @param $type
+     * @param $typeId
+     * @return bool
+     */
+    public function onDeleteEntity($type, $typeId): bool
     {
         $components = $this->em->getRepository(Component::class)->findBy(['type' => $type, 'typeId' => $typeId]);
         foreach ($components as $component) {
             $this->em->remove($component);
         }
+
+        return true;
     }
 
-    public function cloneTemplateBuilder($type, $typeId, $templateId)
+    /**
+     * @param $type
+     * @param $typeId
+     * @param $templateId
+     * @return bool
+     */
+    public function cloneTemplateBuilder($type, $typeId, $templateId): bool
     {
         $componentsOfTemplate = $this->em->getRepository(Component::class)->findBy(['type' => 'BuilderTemplate', 'typeId' => $templateId, 'parentComponent' => null, 'isTemp' => false]);
 
@@ -262,9 +304,17 @@ class Builder
             $this->cloneComponent($c, null, $type, $typeId, $nextPos);
             $nextPos++;
         }
+
+        return true;
     }
 
-    public function makeTemplate($type, $typeId, $templateTitle)
+    /**
+     * @param $type
+     * @param $typeId
+     * @param $templateTitle
+     * @return bool
+     */
+    public function makeTemplate($type, $typeId, $templateTitle): bool
     {
         $builderTemplate = new BuilderTemplate();
         $builderTemplate->setTitle($templateTitle);
@@ -277,5 +327,7 @@ class Builder
             /** @var Component $c */
             $this->cloneComponent($c, null, 'BuilderTemplate', $builderTemplate->getId(), null, true);
         }
+
+        return true;
     }
 }
