@@ -3,37 +3,43 @@
 namespace Akyos\BuilderBundle\Controller;
 
 use Akyos\BuilderBundle\Entity\Component;
+use Akyos\BuilderBundle\Entity\ComponentField;
 use Doctrine\ORM\EntityManagerInterface;
-use Psr\Container\ContainerInterface;
-use Symfony\Bundle\FrameworkBundle\Controller\ControllerTrait;
-use Symfony\Component\HttpFoundation\RequestStack;
-use Symfony\Component\HttpKernel\KernelInterface;
+use Exception;
+use RuntimeException;
+use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\HttpFoundation\Response;
+use Twig\Environment;
 
 
 class RenderComponentController
 {
-    use ControllerTrait;
+    private EntityManagerInterface $entityManager;
+    private ContainerInterface $container;
+    private Environment $twig;
 
-    private $appKernel;
-    private $em;
-    private $request;
-    protected $container;
-
-    public function __construct(KernelInterface $appKernel, EntityManagerInterface $entityManager, ContainerInterface $container, RequestStack $request)
+    public function __construct(EntityManagerInterface $entityManager, ContainerInterface $container, Environment $twig)
     {
-        $this->appKernel = $appKernel;
+        $this->entityManager = $entityManager;
         $this->container = $container;
-        $this->request = $request;
-        $this->em = $entityManager;
+        $this->twig = $twig;
     }
 
+    /**
+     * @param Component $component
+     * @param false $edit
+     * @param null $type
+     * @param null $typeId
+     * @return string|Response
+     * @throws Exception
+     */
     public function renderComponent(Component $component, $edit = false, $type = null, $typeId = null)
     {
+        $em = $this->entityManager;
         $slug = $component->getComponentTemplate()->getSlug();
 
-        $builderClassName = '\Akyos\BuilderBundle\Components\\'.$this->camelCase($slug).'\\'.$this->camelCase($slug).'ComponentController';
-        $appClassName = '\App\Components\\'.$this->camelCase($slug).'\\'.$this->camelCase($slug).'ComponentController';
+        $builderClassName = '\Akyos\BuilderBundle\Components\\'. self::camelCase($slug).'\\'. self::camelCase($slug).'ComponentController';
+        $appClassName = '\App\Components\\'. self::camelCase($slug).'\\'. self::camelCase($slug).'ComponentController';
 
         $params['edit'] = $edit;
         $params['type'] = $type;
@@ -43,13 +49,15 @@ class RenderComponentController
         $params['customClasses'] = $component->getCustomClasses();
         $params['customId'] = $component->getCustomId();
 
-        if(class_exists($appClassName)) {
+        if (class_exists($appClassName)) {
             $view = $appClassName::getTemplateName();
             foreach ($component->getComponentValues() as $value) {
-                if($value->getComponentField()->getType()=== 'entity'){
-                    $params['values'][$value->getComponentField()->getSlug()] = $this->em->getRepository($value->getComponentField()->getEntity())->find((int)$value->getValue());
+                /** @var ComponentField $componentField */
+                $componentField = $value->getComponentField();
+                if($componentField->getType()=== 'entity'){
+                    $params['values'][$componentField->getSlug()] = $em->getRepository($componentField->getEntity())->find((int)$value->getValue());
                 }else{
-                    $params['values'][$value->getComponentField()->getSlug()] = $value->getValue();
+                    $params['values'][$componentField->getSlug()] = $value->getValue();
                 }
             }
 
@@ -58,15 +66,18 @@ class RenderComponentController
                 return $params;
             }
 
-            return $this->renderView('@Components/'.$view, $params);
+            return $this->twig->render('@Components/'.$view, $params);
+        }
 
-        } elseif(class_exists($builderClassName)) {
+        if(class_exists($builderClassName)) {
             $view = $builderClassName::getTemplateName();
             foreach ($component->getComponentValues() as $value) {
-                if($value->getComponentField()->getType()=== 'entity'){
-                    $params['values'][$value->getComponentField()->getSlug()] = $this->em->getRepository($value->getComponentField()->getEntity())->find((int)$value->getValue());
+                /** @var ComponentField $componentField */
+                $componentField = $value->getComponentField();
+                if($componentField->getType()=== 'entity'){
+                    $params['values'][$componentField->getSlug()] = $em->getRepository($componentField->getEntity())->find((int)$value->getValue());
                 }else{
-                    $params['values'][$value->getComponentField()->getSlug()] = $value->getValue();
+                    $params['values'][$componentField->getSlug()] = $value->getValue();
                 }
             }
 
@@ -74,18 +85,26 @@ class RenderComponentController
             if($params instanceof Response) {
                 return $params;
             }
-
-            return $this->renderView($view, $params);
-
-        } else {
-            throw $this->createNotFoundException('Le controller pour ce composant n\'existe pas! : '.$slug);
+            return $this->twig->render($view, $params);
         }
+
+        throw new RuntimeException('Le controller pour ce composant n\'existe pas! : '.$slug);
     }
 
+    /**
+     * @param $componentSlug
+     * @param $values
+     * @param $component
+     * @param $edit
+     * @param $type
+     * @param $typeId
+     * @return string|Response
+     * @throws Exception
+     */
     public function renderComponentBySlug($componentSlug, $values, $component, $edit, $type, $typeId)
     {
-        $builderClassName = '\Akyos\BuilderBundle\Components\\'.$this->camelCase($componentSlug).'\\'.$this->camelCase($componentSlug).'ComponentController';
-        $appClassName = '\App\Components\\'.$this->camelCase($componentSlug).'\\'.$this->camelCase($componentSlug).'ComponentController';
+        $builderClassName = '\Akyos\BuilderBundle\Components\\'. self::camelCase($componentSlug).'\\'. self::camelCase($componentSlug).'ComponentController';
+        $appClassName = '\App\Components\\'. self::camelCase($componentSlug).'\\'. self::camelCase($componentSlug).'ComponentController';
 
         $params['values'] = $values;
         $params['component'] = $component;
@@ -93,27 +112,32 @@ class RenderComponentController
         $params['type'] = $type;
         $params['typeId'] = $typeId;
 
-        if(class_exists($appClassName)) {
+        if (class_exists($appClassName)) {
             $view = $appClassName::getTemplateName();
             $params = $this->container->get('component.'.strtolower($componentSlug))->getParameters($params);
             if($params instanceof Response) {
                 return $params;
             }
-            return $this->renderView('@Components/'.$view, $params);
+            return $this->twig->render('@Components/'.$view, $params);
+        }
 
-        } elseif(class_exists($builderClassName)) {
+        if(class_exists($builderClassName)) {
             $view = $builderClassName::getTemplateName();
             $params = $this->container->get('component.'.strtolower($componentSlug))->getParameters($params);
             if($params instanceof Response) {
                 return $params;
             }
-            return $this->renderView($view, $params);
-
-        } else {
-            throw $this->createNotFoundException('Le controller pour ce composant n\'existe pas!');
+            return $this->twig->render($view, $params);
         }
+
+        throw new RuntimeException('Le controller pour ce composant n\'existe pas!');
     }
 
+    /**
+     * @param $str
+     * @param array $noStrip
+     * @return string|string[]
+     */
     public static function camelCase($str, array $noStrip = [])
     {
         // non-alpha and non-numeric characters become spaces
